CalculateChecksumForPart bug w/ large PartSize
io.LimitedReader.Read does not guarantee that it will fully fill buf in
a single read operation. With a large enough chunksize (e.g., 2 GiB or
greater) io.LimitedReader may require multiple Read operations to fully
populate poolData.

Because m.bufferPool allocates []byte with a capacity of
options.PartSize, it should be safe to remove the use of
io.LimitedReader and instead (a) adjust the size of poolData to the
expected size (up to the final part this is just PartSize, on the last
part it may be a smaller size), and (b) use io.ReadFull to fully fill
poolData from the filehandle.